+++ pytket/tests/predicates_test.py
@@ -53,6 +53,7 @@ from pytket.passes import (  # type: ignore
     RemoveImplicitQubitPermutation,
     FlattenRelabelRegistersPass,
     RoundAngles,
+    PeepholeOptimise2Q,
 )
 from pytket.predicates import (  # type: ignore
     GateSetPredicate,
@@ -864,12 +865,29 @@ def test_flatten_relabel_pass() -> None:
     assert cu.initial_map[Qubit("b", 7)] == Qubit("a", 1)
     assert cu.circuit.qubits == [Qubit("a", 0), Qubit("a", 1)]
 
+    # test default argument
+    c = Circuit()
+    c.add_q_register("p", 4)
+    FlattenRelabelRegistersPass().apply(c)
+    assert all(q.reg_name == "q" for q in c.qubits)
+
 
 def test_round_angles_pass() -> None:
     c0 = Circuit(2).H(0).TK2(0.001, -0.001, 0.001, 0, 1).Rz(0.50001, 0)
     c1 = Circuit(2).H(0).Rz(0.50001, 0)
     assert RoundAngles(8, only_zeros=True).apply(c0)
     assert c0 == c1
+
+
+def test_PeepholeOptimise2Q() -> None:
+    c = Circuit(2).CX(0, 1).CX(1, 0)
+    assert PeepholeOptimise2Q().apply(c)
+    perm = c.implicit_qubit_permutation()
+    assert any(k != v for k, v in perm.items())
+    c = Circuit(2).CX(0, 1).CX(1, 0)
+    assert PeepholeOptimise2Q(allow_swaps=False).apply(c) == False
+    perm = c.implicit_qubit_permutation()
+    assert all(k == v for k, v in perm.items())
 
 
 if __name__ == "__main__":