+++ pytket/docs/changelog.rst
@@ -1,6 +1,18 @@
 Changelog
 =========
 
+1.16.0 (unreleased)
+-------------------
+
+Minor new features:
+
+* Support ``allow_swaps`` parameter for ``PeepholeOptimise2Q``.
+
+Fixes:
+
+* Fix ``FlattenRegisters`` not updating ``ClassicalExpBox``.
+* Fix missing default argument value to ``FlattenRelabelRegistersPass``.
+
 1.15.0 (May 2023)
 -----------------